06.Agents/U05/pyproject.toml
[project]
name = "u05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "google-api-python-client>=2.192.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "psycopg2-binary>=2.9.11",
    "pyngrok>=7.5.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "pytz>=2026.1.post1",
    "twilio>=9.10.3",
]
